pyproject.toml
[project]
name = "deck-doctor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "huggingface-hub>=1.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "transformers>=5.2.0",
    "uvicorn>=0.41.0",
]
